Adds Users, shortener functionality
This commit does it all.  I generated scaffolds for users and sites. I
generated a controller for sessions.  I also generated a basic
controller for the root page.

I added in Omniauth for Github OAuth.  I updated the Users controller to
only allow someone to register or login if their Github username is in
the authorized users YAML file.

The sites model uses `SecureRandom.hex` to generate a random hex string
to use as the default short_code for the shortened links.  If a
collision occurs, it retries 3 more times.  Pretty unlikely since the
only user will be me.

There are two major things left TODO:

1. Write tests
2. User authorization on the controllers.
  * Unauthenticated users should *only* be allowed in the Basic
    Controller.