Fixes issue with unique short codes
When I was writing the migration to add a unique index on short codes
for sites, I mistyped the column name.  This caused SQLite to reject
every new record after the first.  I learned that in SQLite, there is no
way to change a constraint once the table is created.  The best way to
solve this bug seems to be to delete that migration and just change the
table definition from its initial creation.  Luckily, I only have a
single record in the database.

I also added a regression test to keep this from happening again.