Finishes up puma config
This sets the puma webserver config to handle web traffic in a more
production manner.  It adds a few more threads and preloads the
application.  It also sets some things so that doing that doesn't mess
up ActiveRecord.

In addition to that, I added a new capistrano task that will restart the
puma server running on the production server.  Currently, I have a
script that acts as a daemon for the server itself.  On deployment, I
just need to call `rails restart` to restart the puma process.